+++ notebooks/registry/elixirBeacons.html
