+++ scripts/FASPScript6.py
