+++ scripts/FASPScript7.py
